requirements.txt
aiosmtplib     ==   3.0.2
annotated-types ==  0.7.0
anyio         ==    4.4.0
Authlib       ==    1.3.1
bcrypt       ==     4.2.0
certifi      ==     2024.7.4
cffi         ==     1.16.0
click         ==    8.1.7
colorama     ==     0.4.6
cryptography   ==   43.0.0
dnspython      ==   2.6.1
ecdsa         ==    0.19.0
email_validator==   2.2.0
fastapi     ==      0.112.0
greenlet    ==      3.0.3
h11         ==      0.14.0
httpcore     ==     1.0.5
httptools     ==    0.6.1
httpx       ==     0.27.0
idna       ==       3.7
itsdangerous ==     2.2.0
passlib  ==         1.7.4
pip        ==       24.2
pyasn1     ==       0.6.0
pycparser  ==       2.22
pydantic    ==      2.8.2
pydantic_core==     2.20.1
python-dotenv==     1.0.1
python-jose  ==     3.3.0
python-multipart == 0.0.9
PyYAML      ==      6.0.1
rsa          ==     4.9
setuptools    ==    65.5.0
six         ==      1.16.0
sniffio      ==     1.3.1
SQLAlchemy    ==    2.0.31
starlette     ==    0.37.2
typing_extensions== 4.12.2
uvicorn       ==    0.30.5
watchfiles   ==     0.22.0
websockets   ==     12.0